+++ packages/agent/src/index.ts
@@ -18,10 +18,9 @@ export { createGraph } from './agents/graphs/graph.js';
 
 // Agent operators
 export { AgentSelector } from './agents/operators/agentSelector.js';
-export type {
-  AgentConfigResolver,
-  AgentBuilder,
-} from './agents/operators/agentSelector.js';
+
+// Agent types
+export type { AgentConfigResolver, AgentBuilder } from './types/agent.types.js';
 
 // Tool-related exports
 export type {
